command: structure server side for processing commands via websockets
It's not perfect.
Very minimalistic.
Basically takes a command (as JSON text), converts it to JSON via json-c,
sends it to a command handler.
If the reply from the command handler is a non-null JSON object, it gets
sent back via the websocket.

Also, a minimal command is implemented to get the list of camera devices
from a linux machine.

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>